feat(http): added monadic methods to http driver -- Adds monadic interface to @cycle/http
This commit adds new methods to the http driver, named after the request method, to perform http
requests in a monadic style. Rather than having effects encoded as pure instructions returned to a
driver through a stream, effects are encoded as the results of calling a monadic function. They are
performed when the stream resulting from the monadic function is subscribed to.